% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agg}
\alias{agg}
\title{Aggregation matrix}
\usage{
agg(fine_grid, coar_grid)
}
\arguments{
\item{fine_grid}{an \eqn{N_D \times 5} data frame; need more details...}

\item{coar_grid}{an \eqn{M_D \times 2} matrix of coarse-resolution coordinates}
}
\value{
An \eqn{M_D\times N_D} aggregation matrix.
}
\description{
This function generates the aggregation matrix considered in the manuscript.
}
