% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findinit}
\alias{findinit}
\title{Find initial values of GPB parameters}
\usage{
findinit(xxb, mod_param, phi_vals, n_sigmasq)
}
\arguments{
\item{xxb}{a vector of coarse-resolution data of length \eqn{M_D}.}

\item{mod_param}{a list of parameters needed for model fitting. Currently the function requires:
\code{aggmat}: an \eqn{M_D \times N_D} matrix that aggregates fine-resolution data to coarse-resolution data;
\code{bau_distmat}: an \eqn{N_D \times N_D} matrix about pairwise-locaiton distances.}

\item{phi_vals}{a vector of gridded values of GPB range parameter \code{phi}.}

\item{n_sigmasq}{an integer to create a grid of sigmasq.}
}
\value{
The return object is a list that comprises a matrix of log-likelihood values
        and a vector of gridded values of sigmasq
}
\description{
This function finds initial values of GPB range and variance parameters
}
