% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpbau.R
\name{gpbau}
\alias{gpbau}
\title{Function for COS-based statistical downscaling}
\usage{
gpbau(
  coarse_resp,
  coarse_covar,
  prior,
  starting = NULL,
  tuning,
  mod_param,
  mcmc_settings,
  verbose = TRUE
)
}
\arguments{
\item{coarse_resp}{a vector of coarse-resoultion data of length \eqn{M_D}.}

\item{coarse_covar}{an \eqn{M_D \times p} matrix of covariates for the downscaling variable 
of interest. If an intercept is desired, the first column of the matrix should be a vector of ones.}

\item{prior}{a list of priors. Each element of the list corresponds to a combination of an unknown parameter 
and its prior distribution; for example, "\code{phi_unif}" and "\code{sigmasq_invgamma}".}

\item{starting}{a list of starting values. Each element of the list corresponds to an unknown parameter.}

\item{tuning}{a list of tuning parameters. Each element of the list corresponds to an unknown parameter.
The value portion of each element defines the variance of the random walk Metropolis sampler 
proposal distribution for the unknown parameter.}

\item{mod_param}{a list of parameters needed for model fitting. Currently the function requires:
\code{aggmat}: an \eqn{M_D \times N_D} matrix that aggregates fine-resolution data to coarse-resolution data;
\code{bau_distmat}: an \eqn{N_D \times N_D} matrix about pairwise-locaiton distances.}

\item{mcmc_settings}{a list of Markov chain Monte Carlo (MCMC) simulation parameters. 
\code{n_iter}: number of iterations; 
\code{n_burn}: number of burn-in samples;
\code{n_thin}: thinning degree. If \code{verbose = TRUE};
\code{n_report} defines the interval to report MCMC progress and Metropolis acceptance rates.}

\item{verbose}{logical; if true, model specification, MCMC progress, and Metropolis acceptance rates
are printed to the screen.}
}
\value{
An object of class \code{gpbau}. The return object is a list that comprises:

\tabular{ll}{

\code{post_samples} \tab a list of posterior samples. Each element of the list corresponds to an unknown parameter. \cr
\tab \cr
\code{coarse_data} \tab a list that consists of input data 
        (\code{response}, \code{covars}, and \code{coords}). \cr
\tab \cr
\code{mod_spec}: \tab a list that consists of model specifications 
        (\code{mod_param}: parameters needed to fit models; 
         \code{priors}: priors for unknown parameters). \cr
\tab \cr
\code{mcmc_params}: \tab a list that consists of unknown parameters' starting values and Metropolis sampler proposal distribution variances
        used in the MCMC simulation (
        \code{starting}: starting values;
        \code{tuning}: Metropolis sampler proposal distribution variances). \cr
\tab \cr
\code{mh_data}: \tab a list that consists of each unknown parameter's Metropolis 
                     sampler acceptance and rejection indicators.\cr
\tab \cr
\code{runtime}: \tab running time for MCMC simulation calculated using \code{system.time}. \cr

}
}
\description{
This function fits a Gaussian process over basic areal units (GPB) to
             downscale coarse-resolution data via Markov Chain Monte Carlo simulations.
}
