% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpbau.R
\name{predict.gpbau}
\alias{predict.gpbau}
\title{Function for obtaining downscaling samples}
\usage{
\method{predict}{gpbau}(
  object,
  fine_covar,
  type = "independent",
  verbose = TRUE,
  nreport = NULL,
  ...
)
}
\arguments{
\item{object}{an object of class gpbau.}

\item{fine_covar}{an \eqn{N_D \times p} matrix of covariates at the fine resolution.
Each column of the matrix corresponds to the column of the matrix in 
\code{coarse_covar} when it was used to fit GPB.}

\item{type}{A quoted keyword that specifies the type of prediction.
Supported keywrods are \code{"independent"} and \code{"joint"}.}

\item{verbose}{logical; if true, progress of the prediction is printed to the screen.}

\item{nreport}{If \code{verbose = TRUE}, \code{n_report} defines the interval to report
the progress of prediction.}

\item{...}{additional arguments. No additional arguments are supported currently.}
}
\value{
The return object is an \eqn{N_D \times K} matrix, where \eqn{K} is the number 
of posterior predictive samples.
}
\description{
This function produces downscaling samples of the coarse-resolution
variable using GPB via posterior predictive distribution.
}
