% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spds-package.R
\docType{package}
\name{spds}
\alias{spds-package}
\alias{spds}
\title{spds: Spatial-statistical downscaling with uncertainty quantification}
\description{
Downscale coarse-resolution data to fine resolutions using a GPB (Gaussian process over Basic Areal Units). Reference: Zheng, X., Cressie, N., Clarke, D. A., McGeoch, M. A., and Zammit-Mangion, A. (2025). Spatial-statistical downscaling with uncertainty quantification in biodiversity modelling.
}
