% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{thinning}
\alias{thinning}
\title{Thinniing of MCMC samples}
\usage{
thinning(object, nburn, nthin)
}
\arguments{
\item{object}{an object of class gpbau.}

\item{nburn}{an integer: number of burn-in samples.}

\item{nthin}{an integer: thinning degree.}
}
\value{
An object of class \code{gpbau}
}
\description{
This function provides thinning of MCMC samples.
}
